- Cleaned up code in Kernel/*.h with the goal of removing typos and inspection warnings
- Cleaned up all codes to remove compiler warnings; the Trunk now compiles with -Werror.
- Added override where necessary, default where possible, fixed many type/cast issues, typos, etc
- Note, many changes stemmed from renaming timestep to timeStep (in code) or time step (in comments)
- No deprecated functions are used anymore (removed old InfiniteWall::set, get*File(), etc)
- Many deprecated helper functions for setting particle properties have been commented, can be removed if no-one complains.


git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@2953 b3179fc6-a625-46cb-afb8-a2a51a81216d

+++ Drivers/MercuryCG/ConvertP4Files.cpp
@@ -516,7 +516,7 @@ int main(int argc, char *argv[])
 
         unsigned int periodic = 0;
         if (argc>10) {
-            periodic = atoi(argv[10]);
+            periodic = static_cast<unsigned int>(atoi(argv[10]));
             std::cout << "periodic " << periodic << std::endl;
         }
